完成liquidation

+++ programs/bumpin-trade/src/state/pool.rs
@@ -15,7 +15,7 @@ use solana_program::msg;
 pub struct Pool {
     pub pool_key: Pubkey,
     pub pool_mint: Pubkey,
-    pub pool_index: u64,
+    pub pool_index: u16,
     pub pool_mint_vault: Pubkey,
     pub pool_name: [u8; 32],
     pub pool_balance: PoolBalance,
